Add changelog softlink to docs
despite years of issues and pull requests, sphinx still can't handle including a markdown file from a different directory. it also doesn't like a stray dollar sign floating around. so, this commit works around sphinx's shortcomings so we can include the changelog in the docs without having to re-write it in rst and/or create a full duplicate.